server: Implement content-based caching
Caching based on files is limited and doesn't handle the case of slight
changes to a file. Therefore, caching is based on instruction contents
of each translated unit. The configuration and instruction bytes are
combined and hashed. For non-PIC code, the address is includeed as well;
code translated as PIC is reusable independent of the guest address.

This unfortunately duplicates the decoding logic from Rellume, it might
make sense to remove CFG decoding from Rellume in future.