server/cache: Implement write-only mode
To avoid adding another option for this, there is a separate cachemode,
which can be "rw", "r", or "w". This could probably be unified with the
"cache" option as well, but argument parsing is a little messy anyway at
the moment, and caching is still not considered as stable (and therefore
disabled entirely by default).