general: Implement fork
Forking is implemented by first forking the server, sending the socket
fd of the new server to the client, and then forking the client.

The parent-child relationships of the new processes are not optimal: the
forked server is a child of the original server, not the forked client.
This is hard to fix with the current architecture, though.